+++ src/sys/os/browser/web/event/pointer.rs
@@ -4,7 +4,7 @@
 //
 
 use crate::lang::prog::ffi::js::{JsInstant, js_int32, js_number};
-use crate::{KeyMods, WebEventKind};
+use crate::{EventPointerKind, KeyMods, WebEventKind};
 
 #[doc = crate::_tags!(event web)]
 /// A web API Mouse Event.
@@ -20,8 +20,11 @@ use crate::{KeyMods, WebEventKind};
 /// # Compatibility
 ///
 /// Browser Back/Forward mouse buttons are backend-dependent.
-/// Chromium exposes them as DOM mouse events, while Firefox may consume them
-/// for history navigation before the page receives `mousedown`/`mouseup`.
+///
+/// Chromium exposes them as DOM mouse events (`button` 3/4, `buttons` 8/16).
+/// Firefox currently may consume them for history navigation before the page
+/// receives `mousedown`/`mouseup`.
+//
 // WAIT: [firefox-back-forward-buttons](https://bugzilla.mozilla.org/show_bug.cgi?id=1933746)
 #[repr(C)]
 #[derive(Copy, Clone, Debug, PartialEq)]
@@ -64,12 +67,32 @@ impl WebEventMouse {
 /// A web API Pointer Event.
 #[doc = crate::_doc_meta!{
     location("sys/os/browser/web"),
-    test_size_of(WebEventPointer = 48|384; niche Option),
+    test_size_of(WebEventPointer = 56|448; niche Option),
 }]
 ///
 /// Represents a JavaScript pointer event containing relevant properties.
 ///
 /// - <https://developer.mozilla.org/en-US/docs/Web/API/PointerEvent>
+///
+/// # Compatibility
+/// Some Firefox/Linux configurations may report stylus input as
+/// `pointerType = "mouse"` with `pressure = 0.5` and zero tilt.
+///
+/// That is the browser/platform fallback stream.
+//
+// related to MOZ_USE_XINPUT2 in X11:
+// DONE? [pen recognized as mouse in firefox](https://bugzilla.mozilla.org/show_bug.cgi?id=1606832)
+// WAIT: - [Use XInput2 manager on GTK3](https://bugzilla.mozilla.org/show_bug.cgi?id=1207700)
+// WAIT:   - [repeat always false in X11](https://bugzilla.mozilla.org/show_bug.cgi?id=1594003)
+//
+// WAIT: [META stylus/pen handling issues](https://bugzilla.mozilla.org/show_bug.cgi?id=1631377)
+// WAIT: - [spurious pen pointer evens](https://bugzilla.mozilla.org/show_bug.cgi?id=1906968)
+// WAIT: - […incorrect pointerType](https://bugzilla.mozilla.org/show_bug.cgi?id=1449660)
+// WAIT: - […events broken for pen devices](https://bugzilla.mozilla.org/show_bug.cgi?id=1487509)
+// WAIT: - […impl has too many errors](https://bugzilla.mozilla.org/show_bug.cgi?id=1583519)
+// WAIT: - […button state incorrect value](https://bugzilla.mozilla.org/show_bug.cgi?id=1425384)
+// WAIT: - [sActivePointersIds gets 0](https://bugzilla.mozilla.org/show_bug.cgi?id=1904865)
+// WAIT: - [pen input is misrepresented](https://bugzilla.mozilla.org/show_bug.cgi?id=1953665)
 #[repr(C)]
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct WebEventPointer {
@@ -82,6 +105,8 @@ pub struct WebEventPointer {
 
     /// Unique identifier for the pointer device.
     pub id: js_int32, // 4 bytes
+    /// The kind of pointer device.
+    pub kind: EventPointerKind,
     /// The tilt of the stylus along the X-axis (-90° to 90°).
     pub tilt_x: i8, // 1 bytes
     /// The tilt of the stylus along the Y-axis (-90° to 90°).
@@ -105,20 +130,19 @@ impl WebEventPointer {
     /// Returns a new [`WebEventPointer`].
     #[allow(clippy::too_many_arguments)] #[rustfmt::skip]
     pub const fn new(
-        x: js_number,
-        y: js_number,
+        x: js_number, y: js_number,
         pressure: js_number,
         id: js_int32,
-        tilt_x: i8,
-        tilt_y: i8,
+        kind: EventPointerKind,
+        tilt_x: i8, tilt_y: i8,
         twist: u16,
-        button: u8,
-        buttons: u8,
+        button: u8, buttons: u8,
         mods: KeyMods,
         etype: WebEventKind,
         timestamp: JsInstant,
     ) -> Self {
-        Self { x, y, pressure, id, tilt_x, tilt_y, twist, button, buttons, mods, etype, timestamp }
+        Self { x, y, pressure, id, kind, tilt_x, tilt_y, twist,
+            button, buttons, mods, etype, timestamp, }
     }
 }
 
@@ -126,7 +150,8 @@ impl WebEventPointer {
 mod impls {
     use crate::{
         EventButton, EventButtons, EventButtonState, EventKind, EventKindTimed, EventMouse,
-        EventTimestamp, JsInstant, KeyMods, Timed, WebEventKind, WebEventMouse, is, js_number,
+        EventPointer, EventPointerKind, EventTimestamp, JsInstant, KeyMods, Timed, WebEventKind,
+        WebEventMouse, WebEventPointer, f32bits_niche, is, js_int32, js_number,
     };
 
     /* mouse */
@@ -166,6 +191,78 @@ mod impls {
     impl From<Timed<EventMouse, Option<EventTimestamp>>> for WebEventMouse {
         fn from(from: Timed<EventMouse, Option<EventTimestamp>>) -> Self {
             Self::from_event_mouse_timed(from)
+        }
+    }
+
+    /* pointer */
+
+    impl WebEventPointer {
+    /// Converts `WebEventPointer` to `EventKindTimed`.
+    pub const fn to_kind_timed(self) -> EventKindTimed {
+        let kind = EventKind::Pointer(EventPointer::new(
+            self.kind,
+            self.id as u32,
+            self.x as i32, self.y as i32,
+            0, 0,
+            f32bits_niche::new(self.pressure as f32),
+            self.tilt_x, self.tilt_y,
+            self.twist,
+            EventButton::from_web(self.button),
+            EventButtonState::from_web(self.etype),
+            EventButtons::from_bits(self.buttons),
+            self.mods,
+        ));
+        let timestamp = Some(EventTimestamp::from_js(self.timestamp));
+        EventKindTimed::new(kind, timestamp)
+    }
+    /// Converts a timed normalized `EventPointer` back to `WebEventPointer`.
+    pub const fn from_event_pointer_timed(from: Timed<EventPointer, Option<EventTimestamp>>)
+        -> WebEventPointer {
+            let timestamp = is![let Some(t) = from.time, t.to_js(), JsInstant { ms: 0.0 }];
+            WebEventPointer {
+                x: from.value.x as js_number,
+                y: from.value.y as js_number,
+                pressure: from.value.get_pressure() as js_number,
+                id: from.value.id as js_int32,
+                kind: from.value.kind,
+                tilt_x: from.value.tilt_x,
+                tilt_y: from.value.tilt_y,
+                twist: from.value.twist,
+                button: is![let Some(b) = from.value.button, b.to_web(), 255],
+                buttons: from.value.buttons.bits(),
+                mods: from.value.mods,
+                etype: from.value.state.to_web_as_pointer(),
+                timestamp,
+            }
+        }
+    }
+    impl From<WebEventPointer> for EventKindTimed {
+        fn from(from: WebEventPointer) -> Self {
+            from.to_kind_timed()
+        }
+    }
+    impl From<Timed<EventPointer, Option<EventTimestamp>>> for WebEventPointer {
+        fn from(from: Timed<EventPointer, Option<EventTimestamp>>) -> Self {
+            Self::from_event_pointer_timed(from)
+        }
+    }
+
+    impl EventPointerKind {
+        /// Converts a compact web pointer kind into `EventPointerKind`.
+        pub const fn from_web(kind: u8) -> Self {
+            match kind {
+                1 => EventPointerKind::Touch,
+                2 => EventPointerKind::Pen,
+                _ => EventPointerKind::Mouse,
+            }
+        }
+        /// Converts `EventPointerKind` into a compact web pointer kind.
+        pub const fn to_web(self) -> u8 {
+            match self {
+                EventPointerKind::Mouse => 0,
+                EventPointerKind::Touch => 1,
+                EventPointerKind::Pen => 2,
+            }
         }
     }